Enabled Python3.5
Updates to unit tests and code, mostly unicode/bytes differences.
Changded from config to configparser.
