+++ .travis.yml
@@ -1,11 +1,13 @@
 language: python
 python:
   - "2.7"
+  - "3.5"
 # command to install dependencies
 install:
   - pip install tox
 # command to run tests
 script:
   - tox -e pep8
+  - tox -e py35
   - tox -e py27
   - tox -e docs
