+++ opp/tests/functional/api/v1/__init__.py
@@ -3,7 +3,7 @@ import os
 import tempfile
 import unittest
 
-from opp.api import v1 as api
+import opp.flask as flask
 from opp.common import utils
 
 
@@ -15,22 +15,23 @@ class BackendApiTest(unittest.TestCase):
         cls.test_dir = tempfile.mkdtemp(prefix='opp_')
         cls.conf_filepath = os.path.join(cls.test_dir, 'opp.cfg')
         cls.db_filepath = os.path.join(cls.test_dir, 'test.sqlite')
-        with open(cls.conf_filepath, 'wb') as conf_file:
-            conf_file.write("sql_connect: 'sqlite:///%s'" % cls.db_filepath)
+        with open(cls.conf_filepath, 'w') as conf_file:
+            conf_file.write("[DEFAULT]\nsql_connect = sqlite:///%s" %
+                            cls.db_filepath)
             conf_file.flush()
         os.environ['OPP_TOP_CONFIG'] = cls.conf_filepath
         utils.execute("opp-db --config_file %s init" % cls.conf_filepath)
 
         # Create a test client and propgate exceptions to it
-        cls.client = api.app.test_client()
+        cls.client = flask.app.test_client()
         cls.client.testing = True
 
         # Create a user, authenticate and store JWT
         headers = {"Content-Type": "application/json"}
         data = json.dumps({'username': "u", 'password': "p"})
         cls.client.put("/users", headers=headers, data=data)
-        response = cls.client.post("/login", headers=headers, data=data)
-        data = json.loads(response.data)
+        response = cls.client.post("/auth", headers=headers, data=data)
+        data = json.loads(response.data.decode())
         cls.jwt = data['access_token']
 
         # Global headers variable
@@ -54,25 +55,25 @@ class BackendApiTest(unittest.TestCase):
     def _get(self, path, code=200):
         resp = self.client.get(path, headers=self.hdrs)
         self.assertEqual(resp.status_code, code)
-        return json.loads(resp.data)
+        return json.loads(resp.data.decode())
 
     def _put(self, path, data=None, code=200):
         if data:
             data = json.dumps(data)
         resp = self.client.put(path, headers=self.hdrs, data=data)
         self.assertEqual(resp.status_code, code)
-        return json.loads(resp.data)
+        return json.loads(resp.data.decode())
 
     def _post(self, path, data=None, code=200):
         if data:
             data = json.dumps(data)
         resp = self.client.post(path, headers=self.hdrs, data=data)
         self.assertEqual(resp.status_code, code)
-        return json.loads(resp.data)
+        return json.loads(resp.data.decode())
 
     def _delete(self, path, data=None, code=200):
         if data:
             data = json.dumps(data)
         resp = self.client.delete(path, headers=self.hdrs, data=data)
         self.assertEqual(resp.status_code, code)
-        return json.loads(resp.data)
+        return json.loads(resp.data.decode())
