+++ opp/tests/functional/test_dbapi_items.py
@@ -17,8 +17,9 @@ class TestDbApiItems(unittest.TestCase):
         cls.test_dir = tempfile.mkdtemp(prefix='opp_')
         cls.conf_filepath = os.path.join(cls.test_dir, 'opp.cfg')
         cls.db_filepath = os.path.join(cls.test_dir, 'test.sqlite')
-        cls.connection = ("sql_connect: 'sqlite:///%s'" % cls.db_filepath)
-        with open(cls.conf_filepath, 'wb') as conf_file:
+        cls.connection = ("[DEFAULT]\nsql_connect = sqlite:///%s" %
+                          cls.db_filepath)
+        with open(cls.conf_filepath, 'w') as conf_file:
             conf_file.write(cls.connection)
             conf_file.flush()
         utils.execute("opp-db --config_file %s init" % cls.conf_filepath)
