+++ tox.ini
@@ -4,11 +4,12 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py27,pep8
+envlist = py35,py27,pep8
 
 [testenv]
 basepython =
     {py27,pep8,docs}: python2.7
+    {py35}: python3.5
 setenv =
     VIRTUAL_ENV={envdir}
     PYTHONWARNINGS=default::DeprecationWarning
@@ -20,6 +21,10 @@ whitelist_externals =
     bash
     rm
     cp
+
+[testenv:py35]
+commands =
+    pytest opp/tests
 
 [testenv:py27]
 commands =
